+++ paz/applications.py
@@ -28,3 +28,4 @@ from .pipelines import MinimalHandPoseEstimation
 from .pipelines import DetectMinimalHand
 from .pipelines import ClassifyHandClosure
 from .pipelines import SSD512MinimalHandPose
+from .pipelines import SimpleBaselines
