+++ paz/backend/keypoints.py
@@ -488,3 +488,228 @@ def uv_to_vu(keypoints):
         keypoints: Array.
     """
     return keypoints[:, ::-1]
+    
+
+def standardize(data, mean, scale):
+    """Standardize the data.
+    
+    # Arguments
+        data: nxd matrix to normalize
+        mean: Array of means
+        scale: standard deviation
+        
+    # Returns
+        standardized poses2D
+    # """
+    return np.divide((data - mean), scale)
+
+
+def destandardize(data, mean, scale):
+    """Destandardize the data.
+    
+    # Arguments
+        data: nxd matrix to unnormalize
+        mean: Array of means
+        scale: standard deviation
+        
+    # Returns
+        destandardized poses3D
+    """
+    return (data * scale) + mean
+
+
+def initialize_translation(focal_length, joints2D, image_center, ratio):
+    """Computes initial 3D translation of root joint
+    
+    # Arguments
+        focal_length: focal length of the camera in pixels
+        joints2D: 2D root joint from HigherHRNet
+        image_center: center of the image (or principal point)
+        ratio: ration of sum of 3D bones to 2D bones
+        
+    # Returns
+        Array of initial estimate of the global position
+        of the root joint in 3D
+    """
+    z = focal_length * ratio  # depth coord
+    x = (joints2D[:, 0] - image_center[0]) * ratio
+    y = (joints2D[:, 1] - image_center[1]) * ratio
+    translation = np.array((x, y, z))
+    return translation.flatten()
+
+
+def solve_least_squares(solver, compute_joints_distance,
+                        initial_joints_translation, joints3D,
+                        poses2D, focal_length, image_center):
+    """ Solve the least squares
+    
+    # Arguments
+        solver: from scipy.optimize import least_squares
+        compute_joints_distance: global_pose.compute_joints_distance
+        initial_root_translation: initial 3D translation of root joint
+        joints3D: 16 moving joints in 3D
+        poses2d: 2D poses
+        focal_length: focal length
+        img_center: image center
+        
+    Returns
+        optimal translation of root joint for each person
+    """
+    joints_translation = solver(
+        compute_joints_distance, initial_joints_translation, verbose=0,
+        args=(joints3D, poses2D, focal_length, image_center))
+    joints_translation = np.reshape(joints_translation.x, (-1, 3))
+    return joints_translation
+
+
+def project_3D_to_2D(points3D, focal_length, image_center):
+    """ Project points in camera frame from 3D to 2D using intrinsic matrix
+    
+    # Arguments
+        points3D: Nx3 points in camera coordinates (32x3)
+        focal_length: (scalar) Camera focal length
+        img_center: 2x1 image center
+        
+    # Returns
+        Nx2 points in pixel space
+    """
+    z = points3D[:, 2]
+    u = (focal_length / z) * points3D[:, 0] + image_center[0]
+    v = (focal_length / z) * points3D[:, 1] + image_center[1]
+    return np.column_stack((u, v))
+
+
+def get_bones_length(poses2D, poses3D):
+    """Computes sum of bone lengths in 3D
+    
+    #Arguments
+        poses3D: array of predicted poses in 3D (Nx16x3)
+        poses2D: array of poses in 2D    (Nx32)
+        
+    #Returns
+        sum_bones2D: sum of length of all bones in the 3D skeleton
+        sum_bones3D: sum of length of all bones in the 3D skeleton
+    """
+    sum_bones2D = 0
+    sum_bones3D = np.zeros(poses3D.shape[0])
+    start_joints = np.arange(0, 15)
+    end_joints = np.arange(1, 16)
+    for person in poses2D:
+        bone_length = np.linalg.norm(person[start_joints] - person[end_joints])
+        sum_bones2D = sum_bones2D + bone_length
+    for person in poses3D:
+        bone_length = np.linalg.norm(person[start_joints] - person[end_joints])
+        sum_bones3D = sum_bones3D + bone_length
+    return sum_bones2D, sum_bones3D
+    
+
+def compute_reprojection_error(initial_translation, keypoints3D,
+                               keypoints2D, focal_length, image_center):
+    """compute distance between each person joints
+    
+    # Arguments
+        initial_translation: initial guess of position of joint
+        keypoints3D: 3D keypoints to be optimized (Nx16x3)
+        keypoints2D: 2D keypoints (Nx32)
+        focal_length: focal length
+        img_center: principal point of the camera
+        
+    # Returns
+        person_sum: sum of L2 distances between each joint per person
+    """
+    initial_translation = np.reshape(initial_translation, (-1, 3))
+    new_poses3D = np.zeros((keypoints3D.shape))
+    for person in range(len(initial_translation)):
+        new_poses3D[person] = keypoints3D[person] + initial_translation[person]
+    new_poses3D = new_poses3D.reshape((-1, 3))
+    project2D = project_3D_to_2D(new_poses3D, focal_length, image_center)
+    joints_distance = np.linalg.norm(np.ravel(keypoints2D) -
+                                     np.ravel(project2D))
+
+    return np.sum(joints_distance)
+
+
+def merge_into_mean(keypoints2D, args_to_mean):
+    """Rearrange keypoints2D
+
+    # Arguments:
+             keypoints2D: keypoints2D (Nx17x2)
+             args_to_mean: dict of joint indices
+
+    # Returns:
+             keypoints2D: keypoints2D after merging
+            """
+    for point, joints_indices in args_to_mean.items():
+        keypoints2D[:, point] = (keypoints2D[:, joints_indices[0]] + 
+                                 keypoints2D[:,joints_indices[1]]) / 2
+    return keypoints2D
+
+
+def filter_keypoints(keypoints, args_to_joints):
+    """filter keypoints.
+
+    # Arguments
+            keypoints: points in camera coordinates
+            args_to_joints: Array of joints indices
+
+    # Returns
+            filtered keypoints
+    # """
+    return keypoints[:, args_to_joints, :]
+
+
+def filter_keypoints3D(keypoints3D, args_to_joints3D):
+    """Selects 16 moving joints (Neck/Nose excluded) from 32 predicted
+    joints in 3D
+
+    # Arguments
+            keypoints3D: Nx96 points in camera coordinates
+            args_to_joints3D: list of indices
+
+    # Returns
+            filtered_joints_3D: Nx48 points (moving joints)
+    """
+    keypoints_num = len(keypoints3D)
+    keypoints3D = np.reshape(keypoints3D, [keypoints_num, 32, 3])
+    joints3D = filter_keypoints(keypoints3D, args_to_joints3D)
+    return joints3D
+
+
+def filter_keypoints2D(keypoints2D, args_to_mean, h36m_to_coco_joints2D):
+    """Selects 16 moving joints (Neck/Nose excluded) from 17 predicted
+            joints in 2D
+
+    # Arguments
+            keypoints3D: Nx17x2 points in camera coordinates
+            args_to_mean: keypoints indices
+            h36m_to_coco_joints2D: human36m dataset list of joints indices
+
+    # Returns
+            joints2D: Nx32 points (moving joints)
+    """
+    joints2D = filter_keypoints(keypoints2D, h36m_to_coco_joints2D)
+    joints2D = np.reshape(joints2D, [joints2D.shape[0], -1])
+    return joints2D
+
+
+def compute_optimized_pose3D(keypoints3D, joint_translation, focal_length,
+                             image_center):
+    """Compute the optimized 3D pose
+
+    # Arguments
+            keypoints3D: 3D keypoints
+            joint_translation: np array joints translation
+            focal_length: focal_length
+            image_center: image center
+
+    # Returns
+            optimized_poses3D: np array of optimized posed3D
+    """
+    
+    optimized_poses3D = []
+    for person in range(keypoints3D.shape[0]):
+        keypoints3D[person] = keypoints3D[person] + joint_translation[person]
+        points = project_3D_to_2D(keypoints3D[person].reshape((-1, 3)),
+                                  focal_length, image_center)
+        optimized_poses3D.append(np.reshape(points, [1, 64]))
+    return np.array(optimized_poses3D)
