+++ paz/models/keypoint/keypointnet.py
@@ -11,6 +11,14 @@ from tensorflow.keras.models import Model
 from ..layers import ExpectedDepth
 from ..layers import ExpectedValue2D
 from ..layers import SubtractScalar
+from tensorflow.keras.models import Model
+from tensorflow.keras.initializers import HeNormal
+from tensorflow.keras.constraints import MaxNorm
+from tensorflow.keras.layers import (
+    ReLU,
+    Dense,
+    Dropout
+)
 
 
 def block(x, num_filters, dilation_rate, alpha, name, kernel_size=(3, 3)):
@@ -19,7 +27,7 @@ def block(x, num_filters, dilation_rate, alpha, name, kernel_size=(3, 3)):
     x = BatchNormalization()(x)
     x = LeakyReLU(alpha)(x)
     return x
-
+    
 
 def KeypointNet2D(input_shape, num_keypoints, filters=64, alpha=0.1):
     """Model for discovering keypoint locations in 2D space, modified from
@@ -131,3 +139,4 @@ def KeypointNetShared(input_shape, num_keypoints, depth, filters, alpha):
         [uv_volume_A, uv_volume_B])
     inputs, outputs = [image_A, image_B], [uvz_points, uv_volumes]
     return Model(inputs, outputs, name='keypointnet-shared')
+
