+++ paz/pipelines/keypoints.py
@@ -14,6 +14,11 @@ from ..backend.image import get_affine_transform, lincolor
 from ..backend.keypoints import flip_keypoints_left_right, uv_to_vu
 from ..datasets import JOINT_CONFIG, FLIP_CONFIG
 
+import numpy as np
+from paz.processors.keypoints import SimpleBaselines3D
+from paz.datasets.human36m import data_mean2D, data_stdev2D, data_mean3D,\
+    data_stdev3D, dim_to_use3D
+
 
 class KeypointNetSharedAugmentation(SequentialProcessor):
     """Wraps ``RenderTwoViews`` as a sequential processor for using it directly
@@ -384,3 +389,31 @@ class DetectMinimalHand(pr.Processor):
             image = self.draw(image, keypoints)
         image = self.draw_boxes(image, boxes2D)
         return self.wrap(image, boxes2D, keypoints2D, keypoints3D)
+
+
+class SimpleBaselines(pr.Processor):
+    def __init__(self, estimate_keypoints_3D, args_to_mean,
+                 h36m_to_coco_joints2D):
+        """
+        # Arguments
+            estimate_keypoints_3D: 3D simple baseline model
+            args_to_mean: keypoints indices
+            h36m_to_coco_joints2D: h36m joints indices
+
+        # Returns
+            wrapped keypoints2D, keypoints3D
+        """
+        self.estimate_keypoints_2D = HigherHRNetHumanPose2D()
+        self.estimate_keypoints_3D = estimate_keypoints_3D
+        self.baseline_model = SimpleBaselines3D(self.estimate_keypoints_3D,
+                                                data_mean2D, data_stdev2D,
+                                                data_mean3D, data_stdev3D,
+                                                dim_to_use3D, args_to_mean,
+                                                h36m_to_coco_joints2D)
+        self.wrap = pr.WrapOutput(['keypoints2D', 'keypoints3D'])
+
+    def call(self, image):
+        inferences2D = self.estimate_keypoints_2D(image)
+        keypoints2D = inferences2D['keypoints']
+        keypoints2D, keypoints3D = self.baseline_model(np.array(keypoints2D))
+        return self.wrap(keypoints2D, keypoints3D)
