+++ paz/processors/keypoints.py
@@ -11,6 +11,9 @@ from ..backend.keypoints import normalize_keypoints
 from ..backend.keypoints import denormalize_keypoints
 from ..backend.keypoints import compute_orientation_vector
 from ..backend.image import get_scaling_factor
+from ..backend.standard import append_values, predict
+from paz.backend.keypoints import standardize, filter_keypoints2D,\
+    destandardize, merge_into_mean
 
 
 class ProjectKeypoints(Processor):
@@ -218,3 +221,103 @@ class ComputeOrientationVector(Processor):
     def call(self, keypoints):
         orientation = compute_orientation_vector(keypoints, self.parents)
         return orientation
+
+
+
+class FilterKeypoints2D(Processor):
+    def __init__(self, args_to_mean, h36m_to_coco_joints2D):
+        """ Processor class for the filter_keypoints2D in human36m
+        
+        # Arguments
+            args_to_mean: keypoints indices
+            h36m_to_coco_joints2D: h36m joints indices
+
+        # Returns
+            Filtered keypoints2D
+        """
+        super(FilterKeypoints2D, self).__init__()
+        self.h36m_to_coco_joints2D = h36m_to_coco_joints2D
+        self.args_to_mean = args_to_mean
+
+    def call(self, keypoints2D):
+        keypoints2D = merge_into_mean(keypoints2D, self.args_to_mean)
+        return filter_keypoints2D(keypoints2D, self.args_to_mean,
+                                  self.h36m_to_coco_joints2D)
+
+
+class StandardizeKeypoints2D(Processor):
+    def __init__(self, data_mean2D, data_stdev2D):
+        """ Processor class for standerize
+        
+        # Arguments
+            data_mean2D: mean 2D
+            data_stdev2D: standard deviation 2D
+            
+        # Return
+            standerized keypoints2D
+        """
+        self.mean = data_mean2D
+        self.stdev = data_stdev2D
+        super(StandardizeKeypoints2D, self).__init__()
+
+    def call(self, keypoints2D):
+        return standardize(keypoints2D, self.mean, self.stdev)
+
+
+class UnnormalizeData(Processor):
+    def __init__(self, data_mean3D, data_stdev3D, dim_to_use):
+        """ Processor class for unnormalize function in human36m.py
+        
+        # Arguments
+            data_mean3D: mean 3D
+            data_stdev3D: standard deviation 3D
+            dim_to_use: dimensions to use
+            
+        # Return
+            Unormalized data
+        """
+        self.mean = data_mean3D
+        self.stdev = data_stdev3D
+        self.valid = dim_to_use
+        super(UnnormalizeData, self).__init__()
+
+    def call(self, keypoints2D):
+        data = keypoints2D.reshape(-1, 48)
+        rearanged_data = np.zeros((len(data), len(self.mean)), dtype=np.float32)
+        rearanged_data[:, self.valid] = data
+        unnormalized_data = destandardize(rearanged_data, self.mean, self.stdev)
+        return unnormalized_data
+
+
+class SimpleBaselines3D(Processor):
+    def __init__(self, model, data_mean2D, data_stdev2D, data_mean3D,
+                 data_stdev3D, dim_to_use3D, args_to_mean,
+                 h36m_to_coco_joints2D):
+        """ Processor class to predict the 3D keypoints
+        
+        # Arguments
+            model: Simplebaseline 3D model
+            data_mean2D: mean 2D
+            data_stdev2D: standard deviation 2D
+            data_mean3D: data mean 3D
+            data_stdev3D: standar deviation 3D
+            dim_to_use: dimensions to use
+            args_to_mean: joints indices
+            h36m_to_coco_joints2D: h36m data joints indices
+            
+        # Return
+            keypoints2D: 2D keypoints
+            keypoints3D: 3D keypoints
+        """
+        super(SimpleBaselines3D, self).__init__()
+        self.model=model
+        self.filter = FilterKeypoints2D(args_to_mean, h36m_to_coco_joints2D)
+        self.preprocess = StandardizeKeypoints2D(data_mean2D, data_stdev2D)
+        self.postprocess = UnnormalizeData(data_mean3D, data_stdev3D,
+                                           dim_to_use3D)
+        
+    def call(self, keypoints2D):
+        keypoints2D = self.filter(keypoints2D)
+        keypoints3D = predict(keypoints2D, self.model, self.preprocess,
+                              self.postprocess)
+        return keypoints2D, keypoints3D
