+++ tests/examples/human_pose_3d_tests.py
@@ -1,11 +1,14 @@
 import pytest
-from pipeline import SimpleBaselines
+from paz.pipelines.keypoints import SimpleBaselines
 import numpy as np
 from paz.backend.image import load_image
 from paz.backend.camera import Camera
-from linear_model import Simple_Baseline
-from keypoints_processors import SolveTranslation3D
+from paz.models.keypoint.simplebaselines import Simple_Baseline
 from scipy.optimize import least_squares
+from paz.backend.keypoints import filter_keypoints3D
+from paz.backend.keypoints import initialize_translation, solve_least_squares,\
+    get_bones_length, compute_reprojection_error,\
+    compute_optimized_pose3D
 
 
 h36m_to_coco_joints2D = [4, 12, 14, 16, 11, 13, 15, 2, 1, 0, 5, 7, 9, 6, 8, 10]
@@ -30,7 +33,7 @@ def get_poses(pipeline, image):
 
 @pytest.fixture
 def model():
-    model = Simple_Baseline(16, 3, 1024, (32,), 2, 1)
+    model = Simple_Baseline((32,), 16, 3, 1024, 2, 1)
     model.load_weights('weights.h5')
     pipeline = SimpleBaselines(model, args_to_mean, h36m_to_coco_joints2D)
     return pipeline
@@ -185,12 +188,26 @@ def test_simple_baselines_multiple_persons(image_with_multiple_persons_A,
                        keypoints3D_multiple_persons)
     image_height, image_width = image_with_multiple_persons_A.shape[:2]
     intrinsics = get_camera_intrinsics(image_height, image_width)
-    solvetranslation_processor = SolveTranslation3D(least_squares, intrinsics,
-                                                    args_to_joints3D)
-    _, _, poses3D = solvetranslation_processor(keypoints['keypoints2D'],
-                                               keypoints['keypoints3D'])
+    focal_length = intrinsics[0]
+    image_center = intrinsics[1]
 
-    assert np.allclose(poses3D[0], optimised_pose_multiple)
+    joints3D = filter_keypoints3D(keypoints['keypoints3D'], args_to_joints3D)
+    root2D = keypoints['keypoints2D'][:, :2]
+    length2D, length3D = get_bones_length(keypoints['keypoints2D'], joints3D)
+    ratio = length3D / length2D
+    initial_joint_translation = initialize_translation(focal_length, root2D,
+                                                       image_center, ratio)
+    joint_translation = solve_least_squares(least_squares,
+                                            compute_reprojection_error,
+                                            initial_joint_translation,
+                                            joints3D, keypoints['keypoints2D'],
+                                            focal_length, image_center)
+    keypoints3D = np.reshape(keypoints['keypoints3D'], (-1, 32, 3))
+    optimized_poses3D = compute_optimized_pose3D(keypoints3D,
+                                                 joint_translation,
+                                                 focal_length, image_center)
+
+    assert np.allclose(optimized_poses3D[0], optimised_pose_multiple)
 
 
 def test_simple_baselines_single_persons(image_with_single_person_B,
@@ -202,8 +219,22 @@ def test_simple_baselines_single_persons(image_with_single_person_B,
     assert np.allclose(keypoints['keypoints3D'], keypoints3D_single_person)
     image_height, image_width = image_with_single_person_B.shape[:2]
     intrinsics = get_camera_intrinsics(image_height, image_width)
-    solvetranslation_processor = SolveTranslation3D(least_squares, intrinsics,
-                                                    args_to_joints3D)
-    _, _, poses3D = solvetranslation_processor(keypoints['keypoints2D'],
-                                               keypoints['keypoints3D'])
-    assert np.allclose(poses3D, optimised_pose_single)
+    focal_length = intrinsics[0]
+    image_center = intrinsics[1]
+
+    joints3D = filter_keypoints3D(keypoints['keypoints3D'], args_to_joints3D)
+    root2D = keypoints['keypoints2D'][:, :2]
+    length2D, length3D = get_bones_length(keypoints['keypoints2D'], joints3D)
+    ratio = length3D / length2D
+    initial_joint_translation = initialize_translation(focal_length, root2D,
+                                                       image_center, ratio)
+    joint_translation = solve_least_squares(least_squares,
+                                            compute_reprojection_error,
+                                            initial_joint_translation,
+                                            joints3D, keypoints['keypoints2D'],
+                                            focal_length, image_center)
+    keypoints3D = np.reshape(keypoints['keypoints3D'], (-1, 32, 3))
+    optimized_poses3D = compute_optimized_pose3D(keypoints3D,
+                                                 joint_translation,
+                                                 focal_length, image_center)
+    assert np.allclose(optimized_poses3D, optimised_pose_single)
